Moved ellipse tests to correct package
